# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Delivery App API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CreateOrderRequest {
  customerName: string;
  customerPhone: string;
  address: string;
  orderDetails: string;
  notes?: string;
}

export interface UpdateOrderRequest {
  status?: string;
  assignedDriverId?: number;
  assignedDriverName?: string;
}

export interface Order {
  id: number;
  customerName: string;
  customerPhone: string;
  address: string;
  orderDetails: string;
  notes?: string;
  status: string;
  assignedDriverId?: number;
  assignedDriverName?: string;
  createdAt: string;
}

export interface AddPhoneRequest {
  phoneNumber: string;
  label?: string;
}

export interface Phone {
  id: number;
  phoneNumber: string;
  label?: string;
  createdAt: string;
}

export interface AddDriverRequest {
  name: string;
  phone?: string;
}

export interface Driver {
  id: number;
  name: string;
  phone?: string;
  createdAt: string;
}

export interface SiteSettings {
  siteName: string;
  siteTagline: string;
  heroTitle: string;
  heroTitleHighlight: string;
  heroDescription: string;
  availabilityText: string;
  formTitle: string;
  formSubtitle: string;
  successMessage: string;
  primaryColor: string;
  whatsappTemplate: string;
  footerText: string;
  adminPin: string;
  logoImage: string;
}

export interface ErrorResponse {
  error: string;
}