# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Delivery App API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: orders
  - name: phones
  - name: settings
  - name: drivers
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: getOrders
      tags: [orders]
      summary: Get all orders
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /orders/{id}:
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order (assign driver, change status)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderRequest"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /phones:
    get:
      operationId: getPhones
      tags: [phones]
      summary: Get all WhatsApp phone numbers
      responses:
        "200":
          description: List of phone numbers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Phone"
    post:
      operationId: addPhone
      tags: [phones]
      summary: Add a WhatsApp phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddPhoneRequest"
      responses:
        "201":
          description: Phone added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Phone"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /phones/{id}:
    delete:
      operationId: deletePhone
      tags: [phones]
      summary: Delete a WhatsApp phone number
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Phone deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Phone"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /drivers:
    get:
      operationId: getDrivers
      tags: [drivers]
      summary: Get all drivers
      responses:
        "200":
          description: List of drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Driver"
    post:
      operationId: addDriver
      tags: [drivers]
      summary: Add a driver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddDriverRequest"
      responses:
        "201":
          description: Driver added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /drivers/{id}:
    delete:
      operationId: deleteDriver
      tags: [drivers]
      summary: Delete a driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Driver deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get site settings
      responses:
        "200":
          description: Site settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
    put:
      operationId: updateSettings
      tags: [settings]
      summary: Update site settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteSettings"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    CreateOrderRequest:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        address:
          type: string
        orderDetails:
          type: string
        notes:
          type: string
      required: [customerName, customerPhone, address, orderDetails]
    UpdateOrderRequest:
      type: object
      properties:
        status:
          type: string
        assignedDriverId:
          type: integer
        assignedDriverName:
          type: string
    Order:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        address:
          type: string
        orderDetails:
          type: string
        notes:
          type: string
        status:
          type: string
        assignedDriverId:
          type: integer
        assignedDriverName:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, customerName, customerPhone, address, orderDetails, status, createdAt]
    AddPhoneRequest:
      type: object
      properties:
        phoneNumber:
          type: string
        label:
          type: string
      required: [phoneNumber]
    Phone:
      type: object
      properties:
        id:
          type: integer
        phoneNumber:
          type: string
        label:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, phoneNumber, createdAt]
    AddDriverRequest:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
      required: [name]
    Driver:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, createdAt]
    SiteSettings:
      type: object
      properties:
        siteName:
          type: string
        siteTagline:
          type: string
        heroTitle:
          type: string
        heroTitleHighlight:
          type: string
        heroDescription:
          type: string
        availabilityText:
          type: string
        formTitle:
          type: string
        formSubtitle:
          type: string
        successMessage:
          type: string
        primaryColor:
          type: string
        whatsappTemplate:
          type: string
        footerText:
          type: string
        adminPin:
          type: string
        logoImage:
          type: string
      required:
        - siteName
        - siteTagline
        - heroTitle
        - heroTitleHighlight
        - heroDescription
        - availabilityText
        - formTitle
        - formSubtitle
        - successMessage
        - primaryColor
        - whatsappTemplate
        - footerText
        - adminPin
        - logoImage
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]