# lib/api-zod/src/generated/types/siteSettings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Delivery App API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSettings {
  siteName: string;
  siteTagline: string;
  heroTitle: string;
  heroTitleHighlight: string;
  heroDescription: string;
  availabilityText: string;
  formTitle: string;
  formSubtitle: string;
  successMessage: string;
  primaryColor: string;
  whatsappTemplate: string;
  footerText: string;
  adminPin: string;
  logoImage: string;
}