# lib/api-zod/src/generated/types/updateOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Delivery App API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateOrderRequest {
  status?: string;
  assignedDriverId?: number;
  assignedDriverName?: string;
}